+++ assets/linux-dl.sh
@@ -1,3 +1,3 @@
 #! /bin/sh
-curl -L -o mangdl.AppImage https://github.com/MangDL/MangDL/releases/download/3.0.1.0/mangdl-linux-x86_64.AppImage
+curl -L -o mangdl.AppImage https://github.com/MangDL/MangDL/releases/download/3.1.0.0/mangdl-linux-x86_64.AppImage
 chmod +x ./mangdl.AppImage