+++ assets/port-winx64.ps1
@@ -1,6 +1,6 @@
 Invoke-Expression ((New-Object System.Net.WebClient).DownloadString('https://community.chocolatey.org/install.ps1'))
 choco install -y curl 7zip.install
-curl 'https://github.com/MangDL/MangDL/releases/download/3.0.1.0/mangdl-winx64.zip' -o mangdl.zip
+curl 'https://github.com/MangDL/MangDL/releases/download/3.1.0.0/mangdl-winx64.zip' -o mangdl.zip
 $FolderName = 'C:\mangdl\'
 if (Test-Path $FolderName) {
     Remove-Item -Force -Recurse $FolderName