+++ mangdl/api/providers/firstkiss.py
@@ -1,5 +1,7 @@
 from ast import literal_eval
 
+from mangdl.api.base import soup
+
 cloudflare = True
 
 scanlator = "1ST Kiss MANGA"
@@ -11,6 +13,6 @@ rch_fn = "setsu"
 def ch_num_fn(soup):
     return literal_eval(".".join(soup.select_one("#wp-manga-current-chap")["value"].split("-")[1:]))
 
-def manga_id_fn(soup) -> str:
-    cdata = soup.select_one("#wp-manga-js-extra").text
+def manga_id_fn(url, **kwargs) -> str:
+    cdata = soup(url).select_one("#wp-manga-js-extra").text
     return literal_eval(cdata[29:-12])["manga_id"]