+++ mangdl/api/providers/lhtranslation.py
@@ -1,0 +1,6 @@
+scanlator = "Lhtranslation"
+base_url = "https://lhtranslation.net"
+template = "wordpress"
+
+rch_fn = "msa"
+src = "data-src"