+++ mangdl/api/providers/mangahub.py
@@ -1,0 +1,3 @@
+scanlator = "mangaHub"
+base_url = "https://mangahub.online"
+template = "wordpress"