+++ mangdl/api/providers/readmanhua.py
@@ -1,0 +1,8 @@
+scanlator = "Read Manhua"
+base_url = "https://readmanhua.net"
+template = "wordpress"
+
+date_format = r"%d %b"
+search_query_string = "post_type=wp-manga&genre[]=rm&op=1"
+rch_fn = "setsu"
+src = "data-src"