+++ mangdl/api/providers/templates/wordpress.py
@@ -11,11 +11,15 @@ from ....utils.utils import dt, sanitize_text
 from ...base import Ch, Downloader, Manga, soup, urel
 
 class ch_num_fn:
-    def breadcrumb(soup):
+    def breadcrumb(soup: BeautifulSoup):
         return literal_eval(soup.select_one("ol.breadcrumb .active").text.split()[-1])
 
     def wmcc(soup):
         return literal_eval(".".join(soup.select_one("#wp-manga-current-chap")["value"].split("-")[1:]))
+
+class manga_id_fn:
+    def setsu(soup) -> str:
+        return soup.select_one("#manga-chapters-holder")["data-id"]
 
 class rch_fn:
     def msa(url: str, base_url: str, **kwargs: Dict[str, Any]):
@@ -35,26 +39,29 @@ class template:
             "base_url": None,
             "cover_src": "src",
             "ch_num_fn": ch_num_fn.breadcrumb,
+            "date_format": r"%B %d, %Y",
+            "manga_id_fn": manga_id_fn.setsu,
             "rch_fn": soup,
             "rch_num_fn": rch_num_fn.tdo,
             "scanlator": None,
             "search_query_string": "post_type=wp-manga",
             "src": "src",
+            "title_fn": None,
             "total_cs": ".c-blog__heading h1",
         }
         self.prov = prov
         for k, v in attr.items():
             setattr(self, k, getattr(prov, k, v))
         rfk = ["manga_id_fn"]
-        if type(self.rch_fn) == str:
+        if isinstance(self.rch_fn, str):
             self.rch_fn = partial(
                 getattr(rch_fn, self.rch_fn),
                 base_url=self.base_url,
-                **{i: getattr(self.prov, i, None) for i in rfk}
+                **{i: getattr(self, i, None) for i in rfk}
             )
         for i in ["ch_num_fn", "rch_num_fn"]:
             gi = getattr(self, i)
-            if type(gi) == str:
+            if isinstance(gi, str):
                 setattr(self, i, getattr(globals()[i], getattr(prov, i, gi), gi))
 
         self.template = import_module(f".{self.prov.template}", "mangdl.api.providers.templates").template
@@ -78,11 +85,12 @@ class template:
 
     def chdls(self, url: str, chs: int=0) -> List[Dict[Union[float, int, None], str]]:
         op = []
-        for c in self.rch_fn(url).select("li.wp-manga-chapter    a"):
-            cch = c["href"]
-            if chs == 2:
-                cch = self.chapter(cch)
-            op.append({self.rch_num_fn(c["href"]): cch})
+        if chs:
+            for c in self.rch_fn(url).select("li.wp-manga-chapter    a"):
+                cch = c["href"]
+                if chs == 2:
+                    cch = self.chapter(cch)
+                op.append({self.rch_num_fn(c["href"]): cch})
         return op
 
     def manga(self, url: str, chs: int=0) -> Manga:
@@ -100,20 +108,18 @@ class template:
         def mst(s: str, pp: Callable[[str], List[Any]]=lambda x: x):
             return pp(sanitize_text(meta.get(s))) if meta.get(s) else []
 
-        title_fn = getattr(self.prov, "manga_title", None)
-
         rd = [sanitize_text(i.text) for i in self.rch_fn(url).select(".chapter-release-date")]
         def dates(idx: int):
             op = "1970-01-01T00:00:00"
             if rd:
                 rop = rd[idx]
                 if rop:
-                    op = dt(rop, r"%B %d, %Y")
+                    op = dt(rop, self.date_format)
             return op
         return Manga(
             url             = url,
             covers          = [ms.select_one(".summary_image img")[self.cover_src]],
-            title           = title_fn(ms) if title_fn else sanitize_text(ms.select_one("div.post-title h1").text),
+            title           = self.title_fn(ms) if self.title_fn else sanitize_text(ms.select_one("div.post-title h1").text),
             alt_titles      = mst("Alternative"),
             author          = mst("Author(s)", lambda x: x.split(",")),
             artist          = mst("Artist(s)", lambda x: x.split(",")),