+++ mangdl/api/providers/ultralight.py
@@ -1,0 +1,5 @@
+scanlator = "UltraLight"
+base_url = "https://readlight.org"
+template = "wordpress"
+
+rch_fn = "setsu"